Changed the network to use floats instead of ints, started working on NeuralNetworkTrainer

+++ InputSynapse.c++
@@ -2,4 +2,4 @@
 
 InputSynapse::InputSynapse() : weight(0), value(0) {}
 
-InputSynapse::InputSynapse(int weight) : weight(weight), value(0) {}
+InputSynapse::InputSynapse(float weight) : weight(weight), value(0) {}
